project/requirements.txt
Django==5.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0 
gunicorn==21.2.0
whitenoise==6.6.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
dj-database-url==2.1.0
